MB-8901 Return 403 when user is not admin
When checking if an existing user is an admin (in order to grant
access to the admin site), we were comparing with the wrong error
message in the scenario where the admin user was not found, and so the
code went to the next `if` statement and returned a 500 error.

To fix this, we need to check for `RecordNotFoundErrorString`
as opposed to `ErrFetchNotFound`.

While I was in here, I added the missing test for this bug, and I fixed
other places where we were returning a 401 instead of a 403.