MB-4704 Add endpoint for TOO to request reweigh
**Description**

This PR adds a GHC API endpoint to allow a TOO to request a reweigh on
a shipment. Some assumptions I made for the validations:
- We should only allow a reweigh request if the shipment is either
Approved or Diversion Requested.
- A reweigh request is denied if a reweigh already exists for this
shipment.
- Since this endpoint creates a reweigh, and we are not modifying any
shipment fields, I assumed we don't need to pass in and check the eTag.
- The handler should return a Reweigh in the payload. I also included
the Shipment in the Reweigh's payload. I'm not sure if we need that.