MB-9009 Show diversion requested on Move Details
Previously, when a TOO requested a diversion on a shipment from the
MTO tab, the shipment would disappear from the Move details page.

This PR updates the logic to include shipments with the "Diversion
Requested" status, and also labels them as such to make it easier for
the TOO to tell the shipments apart.