Start conversion to using per request loggers
We want to have per request logging with per request tracing, but
almost all of our code uses a top level logger initialized once. This
is the first step towards getting the logger from the request context.

As part of this preparation, remove the per package logger interfaces
which added another layer of checks necessary before using the logger.
Now we can assume we use a zap.Logger everywhere.