Convert to stateless services, introducing appconfig
We want our services to be composable, so that one service can call
another. We also want to be able to have a per request trace id
associated with a logger so that we can correlate log messages in a
single request.

Right now, services are initialized with a database connection pool
and a logger.

When using the `database/sql` package, each database
request uses a different connection to the database. Some of our
services start transactions and then want to call other services. The
"sub-service" uses its own connection pool and thus does not have
visibility to the changes made within the transaction.

The same problem exists for logging because services are logging using
the "global" logger and not one initialized per request.

This PR changes ... everything to use a new appconfig.AppConfig
interface that contains a database connection and a logger. This makes
every service stateless, so that all of the necessary config is passed
as the first argument to every function.

This is a massively invasive change, but one that is necessary to
address the issues above.