Switch to gorilla mux
Gorilla mux provides the feature of being able to determine which
route a request is associated with which will be necessary for
recording metrics associated with the route instead of the file.

It also provides a way to only enable a route for a given host

In addition gorilla mux is actively maintained and goji is not.