Update requestedPickupDate and requestedDeliveryDate on update shipment
This is needed as these two fields are optional on the shipment as they
are not set right away. However, during an update if they are not set to
optional in the api the payload will contain a zero value which can
trigger an update and over write a non zero value that has already been
saved. This change makes them nullable so we can check to tell when
converting the payload to a model.