Migrate the migration;
This patch is two different commits being squashed into one. The commit
messages from the different commits are preserved below.

---

Remove the migration for now;

Since this migration has the potential of running in Production, I have
to remove it in order to follow the Truss Migration Guide.

I will add back things back in a later commit.

---

Add back the migration that was removed earlier;

Since I had to run the migrations on each environment following the
Secure Migrations documentation, I had remove these locally in
241c61f29. I removed the Prime Simulator migration before so I could
follow the documentation that we have without skipping any steps. I
think the documentation should be updated a bit, but for now this worked
out okay.