MB-9285 Add move weights to MTO left nav
Currently, the only items in the MTO left nav are for shipments, so the
way we scroll to each shipment is to assign a unique CSS id for each
shipment section based on the shipment ID.

However, for the new "Move weights" item we want to add (or any other
non-shipment items we might add in the future), we can't use that
pattern, so I defined a new `nonShipmentSections` array that we can
iterate through.

Let me know if there's a better way to do this.