Modify CODEOWNERS to reflect PrimeUI caveat;
The Prime Simulator work does not need to be reviewed by Design nor
Frontend reviewers automatically. Previous to this commit, the
CODEOWNERS file would have requested a review to any CSS or SCSS file
changes. Because Prime Simulator components will modify this file
extension, the files need to be moved into their own directories and
added to CODEOWNERS without an owner. There's inline documentation to
address this that I'll duplicate here for posterity.

> README: Because the Prime UI Simulator work is outside the scope of design /
> frontend reviewing CSS changes, we set the following directories to have an
> "empty owner". This revokes the ownership of .css and .scss files in the
> directive above across the whole repository for any directories below for
> src/*/PrimeUI/
> See an example in the official GitHub CODEOWNERS documentation about directoy
> exceptions in the syntax example here:
> https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-code-owners#codeowners-syntax
> Read more about the caveat around negating patterns in the official GitHub
> CODEOWNERS documentation here:
> https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-code-owners#syntax-exceptions