Add tests to reproduce DLH/DOSHUT and DCRT lookup cache bugs
if we have multiple DCRT on a shipment, the cache does bad stuff.
if we have both DLH and DOSHUT on a shipment, the cache does bad stuff.
I also instrumented the lookup to always recalculate and compare to the cache.
this is temporary, just to help me catch problems.