Make demo config same as staging
Previously, before a recent change to `featureFlags.js`, we didn't have
a specific config for the `demo` environment, and so it defaulted to the
development settings.

However, when we added the `demo` config, we didn't enable certain flags
that are required to do things like customer onboarding and working in
the office app.

This PR adds the same flags that are used in staging.