Refactor TableQueue to use defaultHiddenColumns;
This was brought up by @donaldthai [here][gh-comment] that a can be made
to not have a hard-coded removal of the `id` field from the columns. I
updated it so that it can be configured from outside of the component by
passing in the Array of Strings that the TableQueue component is
expecting. By default, the defaultHiddenColumns will contain the
`['id']` Array so that other components don't break. I checked to make
sure that tests were passing appropriately for these changes.

[gh-comment]: https://github.com/transcom/mymove/pull/7478/files/8ca9b677723926dadcdaea0a400dbe49bf405199#r725385566

Co-authored-by: Donald Thai <donald@truss.works>