Match how date strings are used in the Office app;
This work might be better suited for TRA work, but the way the
MilMove application formats dates is spread across various files rather
than being formatted in a single place in various formats. Taking
inspiration from the ServicesCounselingQueue, the AvailableMovesQueue
adds the Time to the date string format that Office users expect to see
in the application.