Improve dev setup experience
**Description**

For those who prefer to use Homebrew instead of Nix, rather than make
them perform a bunch of steps manually, we can automate most of the
setup with a single command.

Here's how the script works:

1. There's a trussworks/fresh-brew repo that has a main script that
only installs the bare minimum anyone wishing to contribute would need:
Homebrew (which also installs Apple's Command Line Tools if they're not
there yet), and Git. This script is idempotent, meaning it can be run
over and over safely. If you already have Homebrew and Git, it will
automatically update them, and alert you to any issues with your
Homebrew installation.

2. The script also allows you to customize it by adding more tools you
want to install, or commands you want to run, on a project by project
basis. To install additional tools that you want to install via
Homebrew, you add a `Brewfile.local` to your project's repo, like the
one I added in this PR. Additional commands and bash scripts go in a
`fresh-brew.local` file, which takes care of all the required dev setup,
such as installing Node and Go and configuring the shell file with the
proper commands.

This means that if we want to update the Node version for everyone, all
we need to do is update the `.node-version` file, and people can run
`make prereqs`, and it will update everything automatically. Same thing
with updating Go. We specify the new version in `.go-version`, and
`make prereqs` will take care of the rest.