Update Docker image to the latest one
I made an update to the `circleci-docker` repo so that it installs the
same version of Node that this repo uses. This is important for another
PR I'm working on that will update the `check-node-version` script to
ensure that the version specified in `.node-version` is the one everyone
is using.