Refactor TableQueue for manualFilters overrides;
This is necessary because the PrimeUI doesn't have server-side filtering
and instead will use the React Table filtering that the library defaults
to. This should not affect other components that use TableQueue as the
manualFilters Prop defaults to the original `true` value that it was
hard-coded for previous to this commit.