Add ds.provides_services_counseling to GROUP BY
This should have no effect, but it sets up for a future migration.
We will be adding a migration that renames the duty_stations table. It
will look something like this:
    ALTER TABLE duty_stations RENAME TO duty_locations;
    CREATE VIEW duty_stations AS SELECT * FROM duty_locations;

Postgres is not quite as good at figuring things out for views
as it is for tables, so when this query looks at the duty_stations
view, we'll get an error because Postgres can't tell that the services
counseling field will not change within any of the groups.