Update PostalCodeToGBLOC for TOO queue
There were a few goofs that I made when creating this table
that I discovered when I tried to actually use it. Fortunately,
since nothing uses it yet, I don't have to worry about backwards
compatibility in this migration, so it's easy to fix it up.

1. I used a singular name: postal_code_to_gbloc. Pop prefers plural
   names. This is not a big deal to work around, but it's easy to fix.
2. Pop requires tables to have an id column. This is a big one, and
   I had to add an id to all 40k rows to fix it ;_;

I'm also creating a view that will be used in the TOO queue query.
It maps every move_id to a GBLOC based on the postal code of its first
shipment. I couldn't figure out how to express this query in Pop,
without writing raw SQL (which I couldn't do because the query needed
to be composable), but we don't use views for this anywhere else, so it
might be better to go with a less surprising approach.