Update TOO queue to contain non-installation moves
We have some new logic for determining the GBLOC of moves for the purposes
of the TXO queues (which does not apply to the Services Counselor queue).
We're supposed to look up the GBLOC based on the postal code of the move's first shipment.

The TOO queue uses orderFetcher.ListOrders.
This function is also used for the Services Counselor queue.
This function only shows moves that have the same GBLOC as the office
user. The query in this function is already fairly complicated, and I
couldn't find a way to shoehorn in the logic to get the "first shipment"
for a move, without resorting to a raw SQL query, so I used a view,
defined in a previous commit.

I also had to make sure that this change didn't break the SC queue.