Truncate postal_code_to_gbloc for e2e tests
Having this table be cleared for server tests, but not
for e2e tests was making life hard.
Most of our test addresses use the zip code 90210, and expect
the move to be in the LKNQ GBLOC.
With the new GBLOC logic, 90210 goes to KKFA instead.
Rather than updating all of our test data, I opted to clear the table
and map 90210 to LKNQ.