Refactor the shared code a bit;
This moves the NTS UUID into a new type called UUIDList. This is done to
make it easier to enforce what these types are. Also, I wanted to remove
a function call that was unnecessary. This patch also adds a new Struct
called sceneOptionsNTS which can be used to customized an NTS Shipment
Scenario as I will be calling at least four of these in the
subScenarioNTSShipments function.