Remove SetTraceID from HandlerContext and cascading changes
The HandlerContext is initialized once and shared across all handlers.
It should not contain data that changes per request. (It should really be
HandlerConfig, but that change will have to wait for another day).

Remove SetTraceID and replace GetTraceID with GetTraceIDFromRequest to
be explicit about where the TraceID is coming from.

Change the tracing middleware to save/restore a UUID instead of a
string and only convert to a string when logging. This is needed for
storing the UUID in the Event.

Also change Event to take an AppContext and TraceID instead of a
Request and HandlerContext so that the service is not tied to a
handler specific item.