Stop deleting postal_code_to_gbloc in e2e tests
This table was only getting deleted with make e2e_test
and not with make e2e_test_docker, and not when running in circleci.
So it seems like deleting it is not the right thing to do.
I was using this as a way to work around all of our test data
using 90210->LKNQ, when 90210 goes to KKFA in the new data.
Instead of updating all the test data, i changed the TOO user over
to the KKFA GBLOC, and made the TOO tests use this user instead of the
TXO/Services Counselor user.