[MB-10779] (hopefully) fix flaky TOO test by speeding up duty location search (#7938)
* Switch from similarity to % operator

This operator uses the index on the duty station name field,
which was ignored by the similarity function.
It is also stricter than similarity about its threshold, so I
had to try a few threshold values before I found one that didn't
break our tests.

* Change threshold back to default

wow lol i misread the threshold of 0.03 as 0.3 and thought
the similarity function was ignoring the threshold and i
had to do shenanigans to make the new thing work like the
old thing. Turns out everything works the same, reading
is just hard.