Add support for running tests with gotestsum
The current go-junit-reporter does not include the log/failure
message in the XML, which makes it harder to see the error in
CircleCI.

gotestsum JUnit XML support does include the error message, but it
runs the tests different. This adds support for gotestsum while trying
to keep support for go-junit-reporter so we can switch back quickly if
we need to.