Create new transportation office for USCG Base Alameda
Base Alameda and NAS Lemoore (both in CA), were using a transportation
office in Jacksonville, FL that was shared with some installations in
Jacksonville.
This doesn't seem right, so I made a dummy transportation office for
Alameda so I could update its GBLOC without breaking the GBLOCs of
the other bases.
I'm leaving NAS Lemoore alone for now. I think it's wrong for it to
share a transportation office with bases in Florida, but it doesn't
provide SC so it shouldn't matter. In the future, we need to comb
through the transportation office data and clean it up more.