Add nullable flavors of primitive types for swagger
These will be used to distinguish between cases
where a value is provided in a payload and is null from
cases where a value is omitted.
In PATCH requests, we want to set things to null if they're
set explicitly in the payload, and not modify them if they
are not present.