Add new columns for duty station rename
For zero downtime migrations, we need to add new columns with
the new names we're going to use and gradually start using them
before getting rid of the old columns.
I'm also adding triggers to keep the new columns in sync with the
old ones.
A future migration will copy all the old values over to the new
columns.