Increase min-width to 120px to accommodate NTS-R;
The previous value of 8 only allowed for a minimum width of 64px. This
new value of 15 allows for a minimum width of 120px which covers the
longest title for a shipment so far, NTS-release.

Documentation on this change can be viewed in the official USWDS
documentation here: https://designsystem.digital.gov/utilities/height-and-width/#minw