Add order.NewDutyStation back so we can make sure it gets filled
I didn't realize that this field was non-nullable. This threw a bit of a
wrench in my migration plan, but I think we're back on track now.
The problem was that once we switch over to using the new version of the
field, we wouldn't have a way to fill in the old column from Go code, which
would cause insertions to orders to fail. Adding the field back to the model
makes it easy to fill both columns in.

So now we're reading from the new field, and writing to both.

The next steps are:
- migration to copy all values from old column to new column
- migration to make the old column nullable
- remove old column from app code
- migration to drop old column