Don't use click force true;
This disables the validation that Cypress uses to check if a user can
click on something. While Cypress debug messages suggest you do this,
this is an anti-pattern when it comes to testing things in Cypress since
a user cannot actually click an input[type=checkbox] because of styling
done by the USWDS. So this actually clicks on the label element instead
which is a sibling of the input element with further checks on the
`for=` attribute. This simulates the way a user would actually click on
things making the tests more aligned with how a user would actually
click on an checkbox in this application.

I paired asynchronously with @Ryan-Koch on this.

Co-authored-by: Ryan Koch <4325613+Ryan-Koch@users.noreply.github.com>