MB-10675 Fix NaN issue with the `MaskedTextField` component
Thank you @duncan-truss for the suggestion and explanation. Summary: had done it the other way previously because the component wasn't properly marking fields as required when touched, but passing in the field props also passed in the `onChange` prop, which seems to conflict with the component's regular behavior. This way lets us just pass the `onBlur` prop along and seems to still mark the field as required as we want.