Get rid of the only method;
I think the previous commit with changes to the inputs around `focus`
and `blur` events should occur when editing a text input have done the
trick to fix the flakiness. I also ran into some quirks with out the
client app is built when it comes to the Cypress tests being run.
Locally, I was seeing a bunch of things happening where various
attributes on HTML elements were not accurate with the tests.

[=> The docs on this are here](https://docs.cypress.io/api/commands/blur#Usage)