Add in Cypress linting for data-selectors;
This is partly to enforce the findings in the 0052 ADR around using
`data-testid` for selectors in Cypress tests. The rule is only set to
**warn** as the Linter repository recommended this since it's not always
possible to select using a `data-*` attribute. But at the very least we
should try and this linter will make things easier to refactor if we
need to in the future.