Use find instead of hard-coding the path;
It looks like `Docker.raw` location has changed. The **Disk utilization
in Docker for Mac** states that this is in a different place compared to
where it was when this was originally written. This uses `find` to start
at the `Data` path to locate the `Docker.raw` file rather than relying
on a strictly hard-coded file path.

=> Official Docker Documentation https://docs.docker.com/desktop/mac/space/