Add new Auditable method to return errors
During the refactoring of the Internal API, the team discovered that
there was no possible way to return an error for the handlers to the
AuditableAppContextFromRequest. This is causing issues where we cannot
properly rollback database transactions if there is an error.

This new refactor introduces a new method called
AuditableAppContextFromRequestWithErrors which has a `handler` callback
function which returns a tuple of the middleware.Responder and an error.
This can then be used by Pop to properly rollback database transactions.

Co-authored-by: Michael Valdes <michael.valde@truss.works>
Co-authored-by: Paul Sackley <66736616+psackley99@users.noreply.github.com>
Co-authored-by: Andrew Hobson <ahobson@truss.works>