Update display names from Swagger;
This was done entirely manual and is not a good idea for us to continue
doing it this way. I had to track down the names of these fields from
Swagger files and from the Models files for GHC and Orders respectively.
We should investigate how to fix this. I would consider this TRA work.