Removing arbitrary counts of the logs;
I think we can add this type of checking once the logging work has
stabilized. Right now this is changing so many times, that it makes
sense to remove them. We could have a new ticket around this for once
PO9 is done to help detect changes to the history log totals. But I
don't believe that this type of count check gives us better tests.