Refactor move history event mapping
I refactored the move history event mapping to be a little more complex;
previously we could only handle a one to one mapping between event names
and events, and now we can match based on a variable subset of fields,
including tableName and action.

I also refactored the tests to be more contained and reduce repetition
across test files.