Add sketch for initial query to include names;
The problem is that origin_duty_location_id and new_duty_location_id
don't include the name. The history log is supposed to include the
human-readable names from the Duty Locations and not the IDs. This is a
sketch because the SELECT statement with the comments causes the server
to fail with a 500 error if I uncomment those two lines.