Get duty location names in the History log;
This patch adds the duty location names in the History log using the
`duty_locations` table rather than the `duty_location_names` table.
There seems to be an issue with this query though because it constantly
updates all history log entries with the latest duty location name
rather than storing that name for each audit history entry. I will have
to circle back to fix this.