Refactor history updateMoveTaskOrderStatusEvent
The event that corresponds to an update to a move's status is now
informed by the available_to_prime_at field. Since at this point we only
expect to track the change to the 'Approved' status, this seems like a
the most reasonable way to do that. A default case is also there and can
be expanded upon if necessary in future work.