Remove the timed wait & use uploads call for wait;
While pairing with @YanZ777, we found that a specific URL was being
fetched but not waited for when looking at Orders. This should help
alleviate any flakiness that was encountered while testing this.
Sometimes that endpoint would fetch before the Edit button rendered
while other times it would fetch after the Edit button rendered.

Co-authored-by: Yan Zhao <yan.zhao@truss.works>