Return error when both search params are passed
i'm still going back and forth about the API here.
The search page we're building won't be able to search for both
things at once, and it doesn't really make sense because locator
is unique.
but it feels kinda capricious to return an error instead of thinking
up a better API that would prevent this.