Add default dev client_cert in devseed data
Since the client_certs table is linked to the users table, when the
users table is truncated in our devseed data population, the
client_certs is also erased. We need to repopulate the dev db with this
default dev client_cert in order for the client_cert middleware to
accept requests made with this cert.