MB-12926 Add more helper funcs for creating PPMShipments
Specifically:
`MakeApprovedPPMShipmentWaitingOnCustomer` - creates a single PPMShipment that has been approved by a counselor and is waiting on the customer to fill in the info for the actual move and upload necessary documents.

`MakeApprovedPPMShipmentWithActualInfo` - creates a single PMShipment that has been approved by a counselor, has some actual move info, and is waiting on the customer to finish filling out info and upload documents.