MB-13177 Add MovingExpense table with base validation logic & tests
There is an existing type called `MovingExpenseType`, which was used for the old `MovingExpenseDocument` model, so we can't use that name in the new type for our new model. I thought about renaming the old one since we're going to get rid of it, but it has a lot of references, so for the sake of getting this model done, I created the new type with a slightly different name (`MovingExpenseReceiptType`). I still used the shorter name at the table level and for the model field itself because I figure when we get rid of the old type, we can rename the new type more easily. If it's preferred though, we can just use the new type name for the field itself too at both the table and model level.