Use force click to maybe help with flaky test
This test was attempting to click on the row in the TXO table
corresponding to a particular move locator.
This is failing a lot on circle but not locally.
I looked at the cypress artifacts on circle and it seems like the element is
visible when the test fails due to being unable to click on it.
so i think cypress's visibility checks are confused and force is probably
warranted.