Fix ServiceCounselingEditShipmentDetails field validation tests
userEvent.type with an empty string doesn't really work, so I changed it
to type a tab character instead, which has the effect of touching the
field and not entering text.
I also had to change the order in which we entered text into the fields.
I don't know why this worked.