Resolve most form validation test errors
Many of these errors were showing up because the fields that we were
checking for validation errors had not been touched yet, and so were
not displaying any validation errors.
The tests were attempting to trigger the errors by clicking submit buttons
but generally the submit buttons are disabled until validations pass, so
nothing was happening.

There are still 9 failing tests to fix.