Add in the bucket name;
This name is constructed from the values that we're passing in the
`transcom/transcom-infrasec-gov-nonato` repository. At the time of this
commit, the value lives in the transcom/transcom-infrasec-gov-nonato#523
Pull Request.

This value can be set in two places either in the repository above or
by the transcom/terraform-aws-app-environment repository's default value
for the `s3_data_warehouse_bucket` variable.