Add in values as comments from the Terraform;
We realized in transcom/terraform-aws-app-environment#215 that while we
define container and task definitions in Terraform, we override them in
the deployment of our ECS scheduled tasks within the
`cmd/ecs-deploy/task_def.go`. This means that rather than setting things
in our Terraform, we'll need to define the task and container in the Go
binary.