Add a storage flag & plumb it through everything;
This lays out the foundation for how we're going to use this storage
option flag. For now it's only good for `database-export` as an option
since I don't want to have to configure things like how much storage I
give it and what the path is from all the way in the CircleCI config.
That sort of detail can be defined within the task definitions gated by
expressions matching the storage flag. I also removed the option to set
the ephemeral storage size since the minimum you can set is 21 GiB to
override the default. I'm not sure what the default is, but I'm also
sure that I don't need 21 GiB for this database_scratch volume. We also
have alerting if the tasks fails like if the ephemeral storage is too
small.