Add in the VOLUME directive to Dockerfiles;
According to the official AWS documentation, the VOLUME directive needs
to be exposed in order to use the Bind Mounts. Since we reuse this
Dockerfile across all over our scheduled tasks, this might cause issues,
but hopefully it does not since there will only be a Bind Mount added
for the `database-export` storage option.

=> Docs https://docs.aws.amazon.com/AmazonECS/latest/developerguide/bind-mounts.html#bind-mount-considerations