OK well, set the permissions to the user we use;
Learning this now, but I realize that we're running our ECS container
tasks as a user named `1042` which seems arbitrarily set in this commit
here 19523757f6059ea70aeb8c936038c573c9289fae. This means our ECS tasks
run as a non-root user. Also, when we set Bind Mounts for ECS AWS
decides that these are owned by Root. So there's the problem that we
can't access these Bind Mounts. According to AWS support in GitHub,
we've can achieve this with some workarounds. The community around this
has a working solution here in this GitHub comment:
https://github.com/aws/containers-roadmap/issues/736#issuecomment-1124118127

This is hopefully part one of two to get this working. But I feel
confident this will work even if it's consider "ugly".