Reverting the unnecessary cut command;
After running through the Bash in the CircleCI task and locally, I don't
believe the `cut` command here is necessary at all. And to add worse
news, the `cut` command was silently passing because there were no
spaces in our Task Definition nor in the ARN output from AWS. But in my
latest patch before this, there was a space so it failed. This reverts
changes introduced in d7299b5183f.