Import react-router-dom instead of react-router
I was mocking some objects on react-router-dom in my tests, but importing
react-router in my component.
So the mocked things were not being used.
Making the imports match up fixed the test.
Apparently react-router-dom is included in react-router, which also
includes some stuff that we don't need here, so it's better to import
the smaller one.