Adding the functionality and the button to page;
I decided to add the button to **Cancel Shipment** on the _Update
Shipment_ page since that page is used currently to update a single
shipment. The styles have not been worked on. This change is purely to
get the necessary functionality and nothing more. I encourage folks to
add on top of this work to improve the UI/UX if necessary.