Remove unnecessary error returns
error handling with this PDF stuff is a bit weird.
Mostly the PDF functions don't return an error, they save an error to
the PDF object, so you don't know if anything has gone wrong unless you
check that.
I hardly ran into any PDF error cases during development. And I looked
through the PDF library to see when it gives you errors, and they're
mostly due to bad setup, and not things that are likely to happen once
we've got it working. So trying to shoehorn more idiomatic Go error
handling in here doesn't really help with much. We still check for errors
when we're loading stuff and at the end of generation.