First pass at fixing permissions and lots o debug prints
when users log in for the first time (and perhaps, other times??) they
go through authorizeUnknownUser, which doesnt seem to set up their
permissions and roles.
When users who have logged in before log in, they go through
authorizeKnownUser, which does set up their roles and permissions.
So, it seems like the solution to this missing button thing is probalby
to add in the roles and permissions in authorizeUnknownUser, but I don't
really understand how this fits into the whole scheme and if this could
have unintended consequences.