Clean up error handling a bit
If we're going to handle an error return anyway, we might as well check
if we've got any PDF errors.
It's not critically important to check these before the report is done,
but it could prevent us from wasting time drawing things in a report that
will never be rendered.