Fix subsection page breaking and reformat some fields
The main issue here was caused by a find and replace.
I was using the width of a field to calculate how many times we'd have
to wrap a string to make it fit.
And then i started using width=0 to tell fpdf to make that field
fill the remaining space. And I accidentally used that value in
the wrapping calculation.
I also added a lil check to bump section headings over to the next page
if they're really close to the bottom.