Fix storage facility nil dereferences
There was some questionable nil handling in FormatValuesShipment for
storage facilities.
First of all, if somehow shipment.StorageFacility=nil and
shipment.StorageFacilityID != nil, we would have proceeded to reference
stuff on the nil storageFacility. Not a problem yet, but I fixed that.
The second problem was that we weren't handling the case where
storageFacility.Phone or .Email was nil.