MB-13773 Rename `Upload` and `UploadPayload` yaml shared definitions
`UploadPayload` was being flagged by the spectral linter, which I'd forgotten to run. Since this is what our ideal upload payload should be, I changed it to have the `Upload` name. This is the one being used by internal and ghc. Also added `readOnly` to several fields that should have had at (some of which were also flagged by the linter).
`Upload` got renamed to `UploadWithOmissions` since it allows several fields to be omitted. This is the one used by prime and support. Ideally we'd switch them to use the same definition as internal and ghc, but it might take more work. I'll test it in my next commit.