MB-13773 Rename `DocumentPayload` shared yaml definition to `Document`
`DocumentPayload` was being flagged by the spectral linter, which I'd forgotten to run.
For those reading histories, in my previous commit I mentioned trying to get rid of `UploadWithOmissions`. I tried, but it leads to some errors because the prime api currently doesn't return urls as part of the create payload at all. I don't want to introduce that without verifying that it is ok and that would take time so for now it'll stay as it is.