Removing the unproductive comment;
@jesset-truss and I believe comments like these are unproductive and can
potentially waste precious engineering time. The reasoning here is that
we are not using private PSQL APIs and don't need to concern ourselves
with minor versions of PSQL. Keeping these versions in sync between our
Application code and our Infrastructure code must be done more
explicitly and not rely on text comments to get folks to update this
value. See me, @rogeruiz if you have any issues with this. Thanks!

Co-Authored-By: Jesse Tamburino <jesset@truss.works>