Get rid of unnecessarily created moves in tests
This should speed up the tests a bit, and make them less brittle.
When move ID/move are passed in as part of the MTOShipment in the assertions
to a testdatagen function, a new move is created, and then at the end the
shipment's MoveID is overwritten with the one that was provided in the asserions.
So, this works in the tests, but causes a lot of unnecessary moves to
be created, and creating a move requires creating a lot of other
objects, so there are a lot of unnecessary DB operations that happen.

The move should be passed at the top level of the assertions object to
avoid creating a new one.