MB-14365 Add weight ticket fetcher service object
Most of the logic for the fetcher actually exists in a function called `FetchWeightTicketByIDExcludeDeletedUploads` that lives with the weight ticket updater. The next commit will remove that function and replace its usages with the new fetcher.
The new bit of logic is an auth check to make sure that if we are in the customer app, the user retrieving the data owns the data.