MB-14606 Update `cac-prereqs` script to use fewer hard-coded paths
The `pkcs11-tool` and `pkcs15-tool` can both be found using `command`, which is cleaner than using the separate hard-coded paths for `brew` vs `nix` users.
I'm not sure how to make the `opensc-pkcs11.so` path dynamic for all, but at least for `nix`, we can use the `PKCS11MODULE` env var set in `.envrc`.
Moved the path set for `opensc-pkcs11.so` for `nix` to be first so that if someone has both `nix` and `brew` versions, if they're using `nix` for mymove, we use the `nix` version instead of the `brew` version.