Ignore closeout location filter when in closeout gbloc
For office users in closeout gblocs, when they're looking at PPM closeout
tab, every single move will have the same value in the "closeout location"
column, so it doesn't make sense to search on this column. You'll either get
everything or nothing.
Since we put the closeout gbloc in that column instead of the transportation
office name, like we do for Army/AF moves, we would also have to write a
different query. It doesnt make sense to write a new query for something
that won't give you any useful information.