MB-13827 Update progear weight ticket rule function names to match usage
`checkCreateRequiredFields` -> `checkBaseRequiredFields` because we need to run these checks for updates too. That was another change I made, I added this check function to the list of checks run when we do updates. Also moved this function to be above the other required fields function since it's the base of the two.

`checkRequiredFields` -> `checkAdditionalRequiredFields` because they aren't the only required fields, but also aren't all the required fields (since some are covered in the newly renamed `checkBaseRequiredFields`).